requirements.txt
dnspython>=2.4.0
python-whois>=0.9.4
httpx>=0.27.0
rich>=13.7.0
click>=8.1.0
python-dotenv>=1.0.0
